+++ rocketmq-broker/src/topic/manager/topic_route_info_manager.rs
@@ -34,7 +34,7 @@ use rocketmq_rust::RocketMQTokioMutex;
 use tracing::info;
 use tracing::warn;
 
-use crate::error::BrokerError::MQBrokerError;
+use crate::broker_error::BrokerError::MQBrokerError;
 use crate::out_api::broker_outer_api::BrokerOuterAPI;
 
 const GET_TOPIC_ROUTE_TIMEOUT: u64 = 3000;
